refactor(preview): split comic archive preview internals
